pen program

+++ calculator.py
@@ -1,5 +1,5 @@
-x = float(input("Enter first number : "))
-y = float(input("Enter Second number : "))
+x = int(input("Enter first number : "))
+y = int(input("Enter Second number : "))
 addition = x + y
 subtraction = x - y
 multiplication = x * y
@@ -7,3 +7,8 @@ print("Operation Result :")
 print("Addition :", x + y)
 print("Subtraction :", x - y)
 print("Multiplication :", x * y)
+
+# Number
+# Integer
+# Float
+# Double
